Sparse Arrays(problem-3)
There is a collection of input strings and a collection of query strings. 
For each query string, determine how many times it occurs in the list of input strings.
For example, given input  strings = [ab,ba,cd,ab,ba] and queries [ab,ba,cd], we find 2 instances of ab, 2 of ba and 1 of cd. 
For each query, we add an element to our return array, results [2,2,1].